requirements.txt
//numpy>=1.21.0
//flask>=2.0.0
//flask-cors>=3.0.0
//tensorflow>=2.8.0

flask
flask-cors
tensorflow-cpu
numpy
gunicorn